requirements.txt
jupyter
matplotlib
numba
numpy
pillow
pycairo
pyglet
pyyaml
scipy
taichi
tqdm
